refactor: move socials to own subdirectory
