+++ packages/myst-frontmatter/src/project/types.ts
@@ -9,7 +9,7 @@ import type { ProjectSettings } from '../settings/types.js';
 import type { SiteFrontmatter } from '../site/types.js';
 import { SITE_FRONTMATTER_KEYS } from '../site/types.js';
 import type { ExpandedThebeFrontmatter } from '../thebe/types.js';
-import type { SocialLinks } from '../utils/socialLinks.js';
+import type { SocialLinks } from '../socials/types.js';
 
 export const KNOWN_EXTERNAL_IDENTIFIERS = ['arxiv', 'pmid', 'pmcid', 'zenodo'];
