+++ packages/myst-frontmatter/src/project/validators.ts
@@ -28,7 +28,7 @@ import type { ProjectAndPageFrontmatter, ProjectFrontmatter } from './types.js';
 import { validateProjectAndPageSettings } from '../settings/validators.js';
 import { FRONTMATTER_ALIASES } from '../site/types.js';
 import { validateMathMacroObject } from '../math/validators.js';
-import { validateSocialLinks } from '../utils/socialLinks.js';
+import { validateSocialLinks } from '../socials/validators.js';
 
 function getExternalIdentifierValidator(
   key: string,
